+++ addons/sms/models/__init__.py
@@ -8,6 +8,7 @@ from . import mail_message
 from . import mail_notification
 from . import mail_thread
 from . import mail_thread_phone
+from . import models
 from . import res_partner
 from . import sms_api
 from . import sms_sms
